requirements.txt
ultralytics
watchdog
valohai-utils